requirements.txt
beautifulsoup4
django-imagekit
django-widget-tweaks
Pillow
beautifulsoup4
django-ipware
